pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hak-talent-mapping"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx[http2]>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    "supabase>=2.4.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "structlog>=24.2.0",
    "tenacity>=8.3.0",
    "playwright>=1.44.0",
    "python-dotenv>=1.0.0",
    "rapidfuzz>=3.9.0",
    # Phase 3 — enrichment pipeline (LLM via OpenRouter, embeddings via OpenAI)
    "openai>=1.30.0",
    "pinecone-client>=3.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/hak_talent_mapping"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["hak_talent_mapping"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]
